refactor: extract handle_imports_from_inlined_module_with_context to inlined handler
- Moved massive 339-line handle_imports_from_inlined_module_with_context function
  to handlers/inlined.rs where it logically belongs
- Also moved 24-line helper function is_package_init_reexport
- Updated all 4 call sites to use InlinedHandler::handle_imports_from_inlined_module_with_context
- Function handles complex logic for:
  * Wildcard imports from inlined modules
  * Symbol renaming and tree-shaking integration
  * Wrapper init function handling
  * Package re-export detection
- Reduced main mod.rs by 379 lines total (339 + 24 + comments)
- Perfect logical fit: all inlined module import processing now centralized
